+++ tests/testthat/test-TreeSearch_utilities.R
@@ -4,3 +4,14 @@ test_that("EmptyDataset()", {
   expect_equal(TipLabels(tree), names(ret))
   expect_true(inherits(ret, 'phyDat'))
 })
+
+test_that("DoNothing()", {
+  expect_equal(DoNothing(c(1, 2, 3)), c(1, 2, 3))
+  expect_null(DoNothing())
+  
+  tree <- TreeTools::PectinateTree(8)
+  ret <- EmptyPhyDat(tree)
+  
+  expect_equal(DoNothing(tree), tree)
+  expect_equal(DoNothing(ret), ret)
+})